fix math builtins; organize further
